requirements.txt
flask
identity[flask]
flask-session
python-dotenv
requests
